--pcore, --ecore, --thread command line parameters
--pcore, --ecore, --thread command line parameters

Switches:
[-h|--help            ] this help
[-v|--version         ] version info
[-l|--list            ] list of demo types
[-c|--cpu             ] list of CPU properties
   [--pcore           ] using performance core on hybrid CPU
   [--ecore           ] using efficient core on hybrid CPU
[-d|--demo            ] demo type
[-t|--thread          ] thread index

+++ Misc.h
@@ -43,6 +43,6 @@ void random_wrap(signed long long* random);
 
 inline uint64_t serialized_tsc(void);
 
-void SetThread(int32_t threadindex);
+void SetThread(size_t threadindex);
 
 void AVX512_InstrincTest(void);
